requirements.txt
huggingface_hub
nemo-toolkit
hydra-core
lightning
numpy==1.26.4
lhotse
einops
transformers
sentencepiece
pandas
jiwer
librosa
pyannote.audio
webdataset
datasets
editdistance
IPython
pydub